Resolve "支持单机多分片，提升写入性能"

Closes #272 and #262

See merge request FeatureEngineering/rtidb!324

+++ steps/package_fedb.sh
@@ -27,12 +27,12 @@ sed -i "s/--endpoint=.*/--endpoint=${IP}:9527/g" ${package}/conf/tablet.flags
 sed -i "s/#--zk_cluster=.*/--zk_cluster=${IP}:2181/g" ${package}/conf/tablet.flags
 sed -i "s/#--zk_root_path=.*/--zk_root_path=\/fedb/g" ${package}/conf/tablet.flags
 
-
 cp -r tools ${package}/tools
 rm -rf ${package}/tools/dataImporter || :
 rm -rf ${package}/tools/rtidbCmdUtil || :
+ls -l build/bin/
 cp -r build/bin/rtidb ${package}/bin/fedb
-cp -r build/bin/rtidb_mac ${package}/bin/fedb_mac_cli
+test -e build/bin/rtidb_mac &&  cp -r build/bin/rtidb_mac ${package}/bin/fedb_mac_cli
 cd ${package}/bin
 wget http://pkg.4paradigm.com/rtidb/dev/node_exporter
 wget http://pkg.4paradigm.com/rtidb/metricbeat
